+++ src/main/java/com/pullanner/api/controller/token/TokenController.java
@@ -1,16 +1,16 @@
-package com.pullanner.global.auth.jwt.controller;
+package com.pullanner.api.controller.token;
 
 import static com.pullanner.api.ApiUtil.getResponseEntity;
-import static com.pullanner.global.auth.jwt.utils.TokenUtils.getRefreshTokenIdCookie;
+import static com.pullanner.api.TokenUtils.getRefreshTokenIdCookie;
 
-import com.pullanner.global.auth.jwt.argumentresolver.RefreshTokenId;
-import com.pullanner.global.auth.jwt.dto.AccessTokenResponse;
+import com.pullanner.api.argumentresolver.RefreshTokenId;
+import com.pullanner.api.controller.token.dto.AccessTokenResponse;
 import com.pullanner.global.auth.jwt.exception.HackedTokenException;
 import com.pullanner.global.auth.jwt.exception.InvalidTokenException;
-import com.pullanner.global.auth.jwt.service.AccessTokenService;
+import com.pullanner.api.service.token.AccessTokenService;
 import com.pullanner.api.ApiResponseCode;
 import com.pullanner.api.ApiResponseMessage;
-import com.pullanner.global.auth.jwt.service.RefreshTokenService;
+import com.pullanner.api.service.token.RefreshTokenService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
